+++ NEWS.md
@@ -1,5 +1,7 @@
 # jinjar (development version)
 
+* New `parse_template()` to parse a template once and `render()` it multiple times (#13).
+
 # jinjar 0.1.1
 
 * `path_loader()` now correctly finds template files outside the working directory (#11).